Add files via upload
Committing work from 1/11/18. Will continue working today but this is majority of it. Added add and delete functionality as well as a classes.php file. Eventually will make generic header file for stuff that is in every file. Did some CSS styling, failed to make it 3-column Nav. Will try with that again later. All functionalities are currently in place. Now need to work on making it look better, incorporate two more classes so I can fulfill the documentation requirement, and possibly adding a couple more functionalities to make it more robust.